+++ js/data.js
@@ -1,0 +1,4 @@
+let commentArray = ['Всё отлично!','В целом всё неплохо. Но не всё.','Когда вы делаете фотографию, хорошо бы убирать палец из кадра.', 'В конце концов это просто непрофессионально.','Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.','Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.','Лица у людей на фотке перекошены, как будто их избивают.', 'Как можно было поймать такой неудачный момент?!']
+let nameArray = ['Халид-кашмири','Абдул-хьамид','ВисВах','1усмани','Абдурахьман','Иттлам','Мага','Дин','Иба','До и вар сом','К1ай','1арж','Шемиль']
+
+export{commentArray, nameArray}
